renamed some things, removed stuff, updated dockerfile
