+++ .env.vars
@@ -8,7 +8,6 @@ PRICE_THRESHOLD_FILTER=10
 TOKEN_ADDRESS=0x940181a94A35A4569E4529A3CDfB74e38FD98631
 CONNECTOR_TOKENS_ADDRESSES=0x833589fcd6edb6e08f4c7c32d4f71b54bda02913,0x940181a94A35A4569E4529A3CDfB74e38FD98631,0x50c5725949A6F0c72E6C4a641F24049A917DB0Cb,0x4621b7a9c75199271f773ebd9a499dbd165c3191,0x4200000000000000000000000000000000000006,0xb79dd08ea68a908a97220c76d19a6aa9cbde4376,0xf7a0dd3317535ec4f4d29adf9d620b3d8d5d5069,0xcfa3ef56d303ae4faaba0592388f19d7c3399fb4,0xcb327b99ff831bf8223cced12b1338ff3aa322ff,0x2ae3f1ec7f1f5012cfeab0185bfc7aa3cf0dec22,0xc1cba3fcea344f92d9239c08c0568f6f2f0ee452,0x60a3e35cc302bfa44cb288bc5a4f316fdb1adb42,0xd9aaec86b65d86f6a7b5b1b0c42ffa531710b6ca
 STABLE_TOKEN_ADDRESS=0xd9aAEc86B65D86f6A7B5B1b0c42FFA531710b6CA
-BOT_TICKER_INTERVAL_MINUTES=1
 SUGAR_TOKENS_CACHE_MINUTES=10
 SUGAR_LPS_CACHE_MINUTES=10
 ORACLE_PRICES_CACHE_MINUTES=10
