--- .github/workflows/tests.yml
@@ -1,11 +1,0 @@
-name: Tests
-on: push
-jobs:
-  tests:
-    runs-on: ubuntu-latest
-    name: Run tests
-    steps:
-      - uses: actions/checkout@v4
-
-      - name: Run tests
-        run: docker build --no-cache --target test ./
