+++ Dockerfile
@@ -13,12 +13,11 @@ RUN pip install --no-cache-dir -r requirements.txt
 # Copy the rest of the application code to /app
 COPY . .
 
+#Copy the .env example file to the .env file
+COPY .env.vars .env
+
 # Expose the port that the app runs on
 EXPOSE 5000
 
-# Define environment variables
-ENV PORT=5000
-ENV HOST=0.0.0.0
-
 # Run the FastAPI server with Uvicorn
-CMD ["uvicorn", "bots.api:app", "--host", "0.0.0.0", "--port", "5000"]
+CMD ["uvicorn", "sugar_api.api:app", "--host", "0.0.0.0", "--port", "5000"]
