--- deployment.yaml
@@ -1,71 +1,0 @@
-apiVersion: apps/v1
-kind: Deployment
-metadata:
-  name: sugar-api-deployment
-  labels:
-    app: sugar-api
-spec:
-  replicas: 1
-  selector:
-    matchLabels:
-      app: sugar-api
-  template:
-    metadata:
-      labels:
-        app: sugar-api
-    spec:
-      containers:
-        - name: sugar-api
-          image: hftstudio/sugar:latest
-          ports:
-            - containerPort: 5000
-          env:
-            - name: FORCE_REDEPLOY
-              value: "force-redeploy-$(date +%s)"
-            - name: WEB3_PROVIDER_URI
-              value: "https://mainnet.base.org"
-            - name: PROTOCOL_NAME
-              value: "Aerodrome"
-            - name: APP_BASE_URL
-              value: "https://aerodrome.finance"
-            - name: LP_SUGAR_ADDRESS
-              value: "0x51f290CCCD6a54Af00b38edDd59212dE068B8A4b"
-            - name: PRICE_ORACLE_ADDRESS
-              value: "0x3B06c787711ecb5624cE65AC8F26cde10831eb0C"
-            - name: PRICE_BATCH_SIZE
-              value: "40"
-            - name: PRICE_THRESHOLD_FILTER
-              value: "10"
-            - name: TOKEN_ADDRESS
-              value: "0x940181a94A35A4569E4529A3CDfB74e38FD98631"
-            - name: CONNECTOR_TOKENS_ADDRESSES
-              value: "0x833589fcd6edb6e08f4c7c32d4f71b54bda02913,0x940181a94A35A4569E4529A3CDfB74e38FD98631,0x50c5725949A6F0c72E6C4a641F24049A917DB0Cb,0x4621b7a9c75199271f773ebd9a499dbd165c3191,0x4200000000000000000000000000000000000006,0xb79dd08ea68a908a97220c76d19a6aa9cbde4376,0xf7a0dd3317535ec4f4d29adf9d620b3d8d5d5069,0xcfa3ef56d303ae4faaba0592388f19d7c3399fb4,0xcb327b99ff831bf8223cced12b1338ff3aa322ff,0x2ae3f1ec7f1f5012cfeab0185bfc7aa3cf0dec22,0xc1cba3fcea344f92d9239c08c0568f6f2f0ee452,0x60a3e35cc302bfa44cb288bc5a4f316fdb1adb42,0xd9aaec86b65d86f6a7b5b1b0c42ffa531710b6ca"
-            - name: STABLE_TOKEN_ADDRESS
-              value: "0xd9aAEc86B65D86f6A7B5B1b0c42FFA531710b6CA"
-            - name: ADDITIONAL_TOKEN_ADDRESSES
-              value: "0xcbb7c0000ab88b473b1f5afd9ef808440eed33bf"
-            - name: BOT_TICKER_INTERVAL_MINUTES
-              value: "1"
-            - name: SUGAR_TOKENS_CACHE_MINUTES
-              value: "10"
-            - name: SUGAR_LPS_CACHE_MINUTES
-              value: "10"
-            - name: ORACLE_PRICES_CACHE_MINUTES
-              value: "10"
-            - name: UI_POOL_STATS_THUMBNAIL
-              value: "https://i.imgur.com/Wo4vbqN.png"
-      imagePullSecrets:
-        - name: regcred
----
-apiVersion: v1
-kind: Service
-metadata:
-  name: sugar-api-service
-spec:
-  selector:
-    app: sugar-api
-  ports:
-    - protocol: TCP
-      port: 80
-      targetPort: 5000
-  type: LoadBalancer
