+++ requirements.txt
@@ -5,3 +5,8 @@ python-dotenv==1.0.0
 thefuzz==0.19.0
 fastapi==0.95.0
 uvicorn==0.25.0
+# Development dependencies
+pytest==7.4.3
+pytest-asyncio==0.21.1
+black==24.4.2
+flake8==6.1.0
