+++ run.py
@@ -1,4 +1,4 @@
 import uvicorn
 
 if __name__ == "__main__":
-    uvicorn.run("bots.api:app", host="0.0.0.0", port=5000, workers=4)
+    uvicorn.run("sugar_api.api:app", host="0.0.0.0", port=5000, workers=4)
